third_party/WebKit/LayoutTests/fast/media/mq-color-index-01.html

<!DOCTYPE html>
<html>
<head>
  <title>This test validates that the (color-index) media query feature is supported.</title>
  <link rel="author" title="Kenneth Christiansen" href="kenneth.r.christiansen@intel.com"/>
  <link rel="help" href="http://www.w3.org/TR/css3-mediaqueries/#color-index"/>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
  <script>
    test(function(){
      assert_equals(matchMedia('(min-color-index: 1)').matches,
                    !matchMedia('(color-index: 0)').matches);

      assert_equals(matchMedia('all and (color-index)').matches,
                    !matchMedia('not all and (color-index)').matches);
    }, "('color-index') sanity check");
  </script>
<head>
</html>
